+++ bin/inx-chronicle/src/api/stardust/analytics/mod.rs
@@ -1,0 +1,7 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod responses;
+mod routes;
+
+pub use routes::routes;
