+++ bin/inx-chronicle/src/api/stardust/v2/responses.rs
@@ -116,16 +116,3 @@ pub struct MilestoneResponse {
 }
 
 impl_success_response!(MilestoneResponse);
-
-/// Response of `GET /api/v2/analytics/addresses[?start_timestamp=<u32>&end_timestamp=<u32>]`.
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct AddressAnalyticsResponse {
-    #[serde(rename = "totalAddresses")]
-    pub total_addresses: u64,
-    #[serde(rename = "receivingAddresses")]
-    pub recv_addresses: u64,
-    #[serde(rename = "sendingAddresses")]
-    pub send_addresses: u64,
-}
-
-impl_success_response!(AddressAnalyticsResponse);
